@@ -5,7 +5,7 @@
 
     <groupId>de.adorsys.multibanking</groupId>
     <artifactId>multibanking</artifactId>
-    <version>4.0.6-SNAPSHOT</version>
+    <version>4.1.4-SNAPSHOT</version>
     <packaging>pom</packaging>
     <name>Multibanking</name>
     <url>https://github.com/adorsys/multibanking</url>
@@ -32,7 +32,7 @@
     </developers>
 
     <properties>
-        <spring-boot.version>2.0.6.RELEASE</spring-boot.version>
+        <spring-boot.version>2.1.0.RELEASE</spring-boot.version>
         <swagger.version>2.8.0</swagger.version>
         <java.version>1.8</java.version>
         <maven.compiler.source>${java.version}</maven.compiler.source>
@@ -95,11 +95,6 @@
                 <groupId>de.adorsys.envutils</groupId>
                 <artifactId>envutils4j</artifactId>
                 <version>0.2</version>
-            </dependency>
-            <dependency>
-                <groupId>com.github.fakemongo</groupId>
-                <artifactId>fongo</artifactId>
-                <version>2.0.12</version>
             </dependency>
 
             <!-- smartanalytics -->
