@@ -2,6 +2,6 @@ import { Environment } from "./env.interface";
 
 export const ENV: Environment = {
   auth_url: 'http://localhost:8080/auth',
-  api_url: 'http://localhost:8081/api/v1',
-  smartanalytics_url: 'http://localhost:8081/api/v1'
+  api_url: 'http://localhost:8088/api/v1',
+  smartanalytics_url: 'http://localhost:8088/api/v1'
 }
