@@ -134,7 +134,6 @@ export class AnalyticsPage {
     this.periods = [];
     this.lineChartLabels = [];
 
-
     if (referenceGroup) {
       let analyticsStart: Moment = moment().subtract(12, "month").day(1);
       let analyticsEnd: Moment = moment().add(13, "month").day(1);
@@ -225,14 +224,18 @@ export class AnalyticsPage {
     return budget;
   }
 
-  getPeriodAmount(group: BookingGroup, referenceDate: Moment): number {
-    let period: BookingPeriod = this.findPeriod(group.bookingPeriods, referenceDate);
+  getPeriodAmount(group: BookingGroup, periodStart: Moment): number {
+    let period: BookingPeriod = this.findPeriod(group.bookingPeriods, periodStart);
 
     if (period && period.amount) {
       return period.amount;
     }
 
-    if (group.amount) {
+    if (periodStart.isSame(moment(), "month")) {
+        console.log();
+    }
+
+    if (group.amount && moment().isSameOrBefore(periodStart, 'day')) {
       return group.amount;
     }
 
@@ -270,7 +273,7 @@ export class AnalyticsPage {
       return periods.find((period: BookingPeriod) => {
         let periodStart: Moment = moment(period.start);
         let periodEnd: Moment = moment(period.end);
-        return referenceDate.isSameOrAfter(periodStart) && referenceDate.isSameOrBefore(periodEnd);
+        return referenceDate.isSameOrAfter(periodStart, 'day') && referenceDate.isSameOrBefore(periodEnd, 'day');
       });
     }
   }
@@ -351,7 +354,7 @@ export class AnalyticsPage {
     this.forecast = date.isSameOrAfter(moment(), "month");
 
     let period = this.periods.find((period: Moment[]) => {
-      return date.isSameOrAfter(period[0]) && date.isSameOrBefore(period[1]);
+      return date.isSameOrAfter(period[0], 'day') && date.isSameOrBefore(period[1], 'day');
     });
 
     this.budget = this.calculateBudget(period[0], period[1]);
@@ -397,7 +400,7 @@ export class AnalyticsPage {
 
   itemSelected(label: string, aggregatedGroups: AggregatedGroups) {
     let period = this.periods.find((period: Moment[]) => {
-      return this.referenceDate.isSameOrAfter(period[0]) && this.referenceDate.isSameOrBefore(period[1]);
+      return this.referenceDate.isSameOrAfter(period[0], 'day') && this.referenceDate.isSameOrBefore(period[1], 'day');
     });
 
     this.navCtrl.push(BookingGroupPage,
